// web/src/api/models/ResponsesWidgetData.ts
/* tslint:disable */
/* eslint-disable */
/**
 * mindscape
 * This is the swagger page for the Project mindscape generated with Egg-go. use this to test your database connection
 *
 * The version of the OpenAPI document: 0.0.1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { SchemasWidgetLayout } from './SchemasWidgetLayout';
import {
	SchemasWidgetLayoutFromJSON,
	SchemasWidgetLayoutFromJSONTyped,
	SchemasWidgetLayoutToJSON,
	SchemasWidgetLayoutToJSONTyped,
} from './SchemasWidgetLayout';
import type { SchemasWidgetProperty } from './SchemasWidgetProperty';
import {
	SchemasWidgetPropertyFromJSON,
	SchemasWidgetPropertyFromJSONTyped,
	SchemasWidgetPropertyToJSON,
	SchemasWidgetPropertyToJSONTyped,
} from './SchemasWidgetProperty';

/**
 *
 * @export
 * @interface ResponsesWidgetData
 */
export interface ResponsesWidgetData {
	/**
	 *
	 * @type {string}
	 * @memberof ResponsesWidgetData
	 */
	id?: string;
	/**
	 *
	 * @type {SchemasWidgetLayout}
	 * @memberof ResponsesWidgetData
	 */
	layout?: SchemasWidgetLayout;
	/**
	 *
	 * @type {{ [key: string]: SchemasWidgetProperty; }}
	 * @memberof ResponsesWidgetData
	 */
	properties?: { [key: string]: SchemasWidgetProperty };
	/**
	 *
	 * @type {Array<string>}
	 * @memberof ResponsesWidgetData
	 */
	required?: Array<string>;
	/**
	 *
	 * @type {string}
	 * @memberof ResponsesWidgetData
	 */
	title?: string;
	/**
	 *
	 * @type {string}
	 * @memberof ResponsesWidgetData
	 */
	type?: string;
}

/**
 * Check if a given object implements the ResponsesWidgetData interface.
 */
export function instanceOfResponsesWidgetData(
	value: object,
): value is ResponsesWidgetData {
	return true;
}

export function ResponsesWidgetDataFromJSON(json: any): ResponsesWidgetData {
	return ResponsesWidgetDataFromJSONTyped(json, false);
}

export function ResponsesWidgetDataFromJSONTyped(
	json: any,
	ignoreDiscriminator: boolean,
): ResponsesWidgetData {
	if (json == null) {
		return json;
	}
	return {
		id: json['id'] == null ? undefined : json['id'],
		layout:
			json['layout'] == null
				? undefined
				: SchemasWidgetLayoutFromJSON(json['layout']),
		properties:
			json['properties'] == null
				? undefined
				: mapValues(json['properties'], SchemasWidgetPropertyFromJSON),
		required: json['required'] == null ? undefined : json['required'],
		title: json['title'] == null ? undefined : json['title'],
		type: json['type'] == null ? undefined : json['type'],
	};
}

export function ResponsesWidgetDataToJSON(json: any): ResponsesWidgetData {
	return ResponsesWidgetDataToJSONTyped(json, false);
}

export function ResponsesWidgetDataToJSONTyped(
	value?: ResponsesWidgetData | null,
	ignoreDiscriminator: boolean = false,
): any {
	if (value == null) {
		return value;
	}

	return {
		id: value['id'],
		layout: SchemasWidgetLayoutToJSON(value['layout']),
		properties:
			value['properties'] == null
				? undefined
				: mapValues(value['properties'], SchemasWidgetPropertyToJSON),
		required: value['required'],
		title: value['title'],
		type: value['type'],
	};
}